// src/pages/Allegro/types/AllegroOffer.ts
export type AllegroOffer = {
  brand: string;
  isPromoted: boolean;
  isSuperSeller: boolean;
  position: number;
  price: number;
  productName: string;
  pageProductName: string;
  productUrl: string;
  url: string;
  seller: string;
  offerId: string;
  unitsSold: number | null;
  imageUrl: string;
  isSuperPrice: boolean;
  hasFreeDelivery: boolean;
};